Backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
pydantic==1.10.7
python-barcode==0.13.1
passlib==1.7.4
PyJWT==2.6.0
pandas==1.4.3
openpyxl==3.0.10
